+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/ResponseParametersFrom.kt
@@ -18,14 +18,6 @@ sealed class ResponseParametersFrom {
         override val hasBeenEncrypted: Boolean = false
     }
 
-    data class JweForJws(
-        val jweDecrypted: at.asitplus.signum.indispensable.josef.JweDecrypted<String>,
-        val parent: ResponseParametersFrom,
-        override val parameters: AuthenticationResponseParameters,
-    ) : ResponseParametersFrom() {
-        override val hasBeenEncrypted: Boolean = true
-    }
-
     data class JweDecrypted(
         val jweDecrypted: at.asitplus.signum.indispensable.josef.JweDecrypted<AuthenticationResponseParameters>,
         val parent: ResponseParametersFrom,