+++ CHANGELOG.md
@@ -27,7 +27,8 @@ Release 5.10.0 (Unreleased):
    - Use correct response encryption acc. to values stated in `encrypted_response_enc_values_supported` inside `client_metadata`
    - Correct evaluation of `vp_formats_supported` inside `client_metadata`
    - Advertise correct wallet metadata for retrieving authn requests, e.g. `client_id_prefixes_supported`
-   - Drop support for signed-then-encrypted responses
+   - Drop support for signed-then-encrypted responses, as in OpenID4VP 1.0
+   - Deprecate method `validateAuthnResponse(input: Map<String, String>)` in `OpenId4VpVerifier` because it eluded correct input parsing
  - SD-JWT:
    - Fix creation of SD JWTs containing structures that are selectivey disclosable